postgres/requirements.txt
psycopg2-binary==2.9.9
tqdm==4.65.0
